Start exporting statement values
Includes some commented-out code that attempted to load the original
data type definitions from Wikibase, but that turned out to be
impossible, they rely on WikibaseRepo in too many places. We’ll have to
hand-craft our own data type definitions. (For now, just 'string' is
available, as it’s the easiest option.)